changed _is_locked to a uint32 and started the bus numbering at 0

+++ src/lib/drivers/device/nuttx/SPI.hpp
@@ -153,7 +153,7 @@ protected:
 
 	LockMode			_locking_mode;	/**< selected locking mode */
 
-	static uint8_t 		_is_locked; /** Bit mask. Bit position corresponds to bus number. */
+	static uint32_t 		_is_locked; /** Bit mask. Bit position corresponds to bus number. */
 
 private:
 	uint32_t			_device;
